requirement.txt
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
Automat==22.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==41.0.5
cssselect==1.2.0
fastapi==0.104.1
filelock==3.13.1
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
lxml==4.9.3
outcome==1.3.0.post0
packaging==23.2
parsel==1.8.1
Protego==0.3.0
psycopg2-binary==2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.0
pydantic_core==2.14.1
PyDispatcher==2.0.7
pyOpenSSL==23.3.0
PySocks==1.7.1
python-dotenv==1.0.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
selenium==4.15.2
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.16
starlette==0.27.0
tldextract==5.1.0
trio==0.23.1
trio-websocket==0.11.1
Twisted==22.10.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
w3lib==2.1.2
wsproto==1.2.0
zope.interface==6.1
